// src/utils/enums/discipline.enum.ts
export enum DisciplineEnum {
  ATHLETICS = 'athletics',
  SWIMMING = 'swimming',
  CYCLING = 'cycling',
  FOOTBALL = 'football',
  BASKETBALL = 'basketball',
  TENNIS = 'tennis',
  VOLLEYBALL = 'volleyball',
  GYMNASTICS = 'gymnastics',
  BOXING = 'boxing',
  JUDO = 'judo',
  WEIGHTLIFTING = 'weightlifting',
  POWERLIFTING = 'powerlifting',
  CALISTHENICS = 'calisthenics',
  CROSSFIT = 'crossfit',
  RUGBY = 'rugby',
  BASEBALL = 'baseball',
  KARATE = 'karate',
  TAEKWONDO = 'taekwondo',
  CLIMBING = 'climbing',
  OTHER = 'other',
}